refactor(core): refactor to move store to separate file and update names
